# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI Simple Project","description":"A simple and straightforward FastAPI application","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","description":"Main route","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/platos/":{"get":{"summary":"Get All Platos","description":"Get all dishes\n\nReturns a list with all available dishes in the menu","operationId":"get_all_platos_platos__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Plato"},"type":"array","title":"Response Get All Platos Platos  Get"}}}}}},"post":{"summary":"Create Plato","description":"Create a new dish\n\n- **name**: Dish name (required)\n- **precio**: Dish price in euros (must be greater than 0)","operationId":"create_plato_platos__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlatoCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Plato"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/platos/{plato_id}":{"get":{"summary":"Get Plato","description":"Get a specific dish by its ID\n\n- **plato_id**: Unique dish ID","operationId":"get_plato_platos__plato_id__get","parameters":[{"name":"plato_id","in":"path","required":true,"schema":{"type":"integer","title":"Plato Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Plato"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Plato","description":"Update an existing dish\n\n- **plato_id**: Unique ID of the dish to update\n- **name**: New dish name (optional)\n- **precio**: New dish price (optional)","operationId":"update_plato_platos__plato_id__put","parameters":[{"name":"plato_id","in":"path","required":true,"schema":{"type":"integer","title":"Plato Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlatoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Plato"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Patch Plato","description":"Partial update of a dish (PUT alias)\n\n- **plato_id**: Unique ID of the dish to update\n- **name**: New dish name (optional)\n- **precio**: New dish price (optional)","operationId":"patch_plato_platos__plato_id__patch","parameters":[{"name":"plato_id","in":"path","required":true,"schema":{"type":"integer","title":"Plato Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlatoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Plato"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Plato","description":"Delete a dish from the menu\n\n- **plato_id**: Unique ID of the dish to delete","operationId":"delete_plato_platos__plato_id__delete","parameters":[{"name":"plato_id","in":"path","required":true,"schema":{"type":"integer","title":"Plato Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/platos/stats/summary":{"get":{"summary":"Get Platos Stats","description":"Get menu statistics\n\nReturns statistical information about available dishes","operationId":"get_platos_stats_platos_stats_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Plato":{"properties":{"id":{"type":"integer","title":"Id","description":"Unique dish ID","example":1},"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name","description":"Dish name","example":"Paella Valenciana"},"precio":{"type":"number","exclusiveMinimum":0.0,"title":"Precio","description":"Dish price in euros","example":15.5}},"type":"object","required":["id","name","precio"],"title":"Plato","description":"Model to represent a dish from the menu","example":{"id":1,"name":"Paella Valenciana","precio":15.5}},"PlatoCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name","description":"Dish name","example":"Paella Valenciana"},"precio":{"type":"number","exclusiveMinimum":0.0,"title":"Precio","description":"Dish price","example":15.5}},"type":"object","required":["name","precio"],"title":"PlatoCreate","description":"Model to create a new dish","example":{"name":"Paella Valenciana","precio":15.5}},"PlatoUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Name","description":"Dish name"},"precio":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"null"}],"title":"Precio","description":"Dish price"}},"type":"object","title":"PlatoUpdate","description":"Model to update an existing dish","example":{"name":"Paella Valenciana Actualizada","precio":18.0}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}